drivers: leds: Fix atomic scheduling bug
This fixes:

> BUG: scheduling while atomic: mm-qcamera-daem/13409/0x00000002

which itself then causes mm-qcamera-daemon to crash.

TORCH is buggy since oreo for us so we disabled it in certain places
but this commit ensures that the wrong scheduling which happens when
the bug occurs does not crash the qcamera daemon anymore.

Hint:
Even when there was no hotkey for torch defined anywhere in android the
above happened.

Force test:
enable power button long press = torch in android settings, lock screen,
press power button for 2 sec or so.

Change-Id: I4ee865260e8fb3954b3d3191c915c461a4a2afc3